+++ src/lib/gene/geneconv.ts
@@ -4,7 +4,7 @@ import { API_GENECONV_URL } from '@/lib/edb/edb'
 import { httpFetch } from '@/lib/http/http-fetch'
 import { capitalCase } from '@/lib/text/capital-case'
 import { QueryClient } from '@tanstack/react-query'
-import { SeriesData } from '../dataframe'
+import type { SeriesData } from '../dataframe'
 import { AnnotationDataFrame } from '../dataframe/annotation-dataframe'
 
 export interface GeneConvResult {