+++ src/lib/text/lines.ts
@@ -25,11 +25,11 @@ export function textToLines(text: string, opts: IToLineOpts = {}): string[] {
   let ret = text.split(/[\r\n]+/g)
 
   if (trim) {
-    ret = ret.map((line) => line.trim())
+    ret = ret.map(line => line.trim())
   }
 
   if (rtrim) {
-    ret = ret.map((line) => line.trimEnd())
+    ret = ret.map(line => line.trimEnd())
   }
 
   if (removeEmptyLines) {
@@ -39,12 +39,12 @@ export function textToLines(text: string, opts: IToLineOpts = {}): string[] {
   if (splitOnPunctuation) {
     const punctRegex = /[,;\|]/g
     ret = ret
-      .map((line) => line.replace(punctRegex, ' '))
-      .map((line) => line.replace(/\s{2,}/g, ' ')) // replace multiple spaces with single space
-      .map((line) => line.split(' '))
+      .map(line => line.replace(punctRegex, ' '))
+      .map(line => line.replace(/\s{2,}/g, ' ')) // replace multiple spaces with single space
+      .map(line => line.split(' '))
       .flat()
-      .map((line) => line.trim())
-      .filter((line) => line.length > 0)
+      .map(line => line.trim())
+      .filter(line => line.length > 0)
   }
 
   return ret
@@ -63,7 +63,7 @@ export function textToTokens(
     ...opts,
   }
 
-  return textToLines(text, { rtrim: true }).map((line) =>
+  return textToLines(text, { rtrim: true }).map(line =>
     line.replaceAll('"', '').split(sep)
   )
 }